+++ modules/bars/stancebar.lua
@@ -21,19 +21,13 @@ local BUTTONS = {
 }
 
 local TOP_POINT = {
-	p = "BOTTOM",
-	anchor = "UIParent",
-	rP = "BOTTOM",
-	x = 0,
-	y = 156,
+	ls = {p = "BOTTOM", anchor = "UIParent", rP = "BOTTOM", x = 0, y = 156},
+	traditional = {p = "BOTTOM", anchor = "UIParent", rP = "BOTTOM", x = 0, y = 156},
 }
 
 local BOTTOM_POINT = {
-	p = "BOTTOM",
-	anchor = "UIParent",
-	rP = "BOTTOM",
-	x = 0,
-	y = 128,
+	ls = {p = "BOTTOM", anchor = "UIParent", rP = "BOTTOM", x = 0, y = 128},
+	traditional = {p = "BOTTOM", anchor = "UIParent", rP = "BOTTOM", x = 0, y = 128},
 }
 
 local LAYOUT = {
@@ -52,7 +46,7 @@ local LAYOUT = {
 }
 
 local function getBarPoint()
-	return LAYOUT[E.PLAYER_CLASS]
+	return LAYOUT[E.PLAYER_CLASS][E.UI_LAYOUT]
 end
 
 local function button_Update(self)