+++ modules/unitframes/units/focustarget.lua
@@ -56,8 +56,8 @@ function UF:CreateFocusTargetFrame(frame)
 
 	local health = self:CreateHealth(frame, textParent)
 	health:SetFrameLevel(level + 1)
-	health:SetPoint("LEFT", frame, "LEFT", 0, 0)
-	health:SetPoint("RIGHT", frame, "RIGHT", 0, 0)
+	health:SetPoint("LEFT", frame.Insets.Left, "RIGHT", 0, 0)
+	health:SetPoint("RIGHT", frame.Insets.Right, "LEFT", 0, 0)
 	health:SetPoint("TOP", frame.Insets.Top, "BOTTOM", 0, 0)
 	health:SetPoint("BOTTOM", frame.Insets.Bottom, "TOP", 0, 0)
 	health:SetClipsChildren(true)
@@ -67,10 +67,6 @@ function UF:CreateFocusTargetFrame(frame)
 
 	local power = self:CreatePower(frame, textParent)
 	power:SetFrameLevel(level + 1)
-	power:SetPoint("LEFT", frame, "LEFT", 0, 0)
-	power:SetPoint("RIGHT", frame, "RIGHT", 0, 0)
-	power:SetPoint("TOP", frame.Insets.Bottom, "TOP", 0, -2)
-	power:SetPoint("BOTTOM", frame.Insets.Bottom, "BOTTOM", 0, 0)
 	frame.Power = power
 
 	power.UpdateContainer = function(_, shouldShow)
@@ -85,6 +81,8 @@ function UF:CreateFocusTargetFrame(frame)
 		end
 	end
 
+	frame.Insets.Bottom:Capture(power, 0, 0, -2, 0)
+
 	frame.Name = self:CreateName(frame, textParent)
 
 	frame.RaidTargetIndicator = self:CreateRaidTargetIndicator(frame, textParent)
@@ -98,7 +96,6 @@ function UF:CreateFocusTargetFrame(frame)
 
 	local border = E:CreateBorder(textureParent)
 	border:SetTexture("Interface\\AddOns\\ls_UI\\assets\\border-thick")
-	border:SetSize(16)
 	border:SetOffset(-6)
 	frame.Border = border