+++ modules/unitframes/units/target.lua
@@ -24,6 +24,7 @@ local function frame_Update(self)
 		self:UpdateInsets()
 		self:UpdateHealth()
 		self:UpdateHealthPrediction()
+		self:UpdatePortrait()
 		self:UpdatePower()
 		self:UpdateCastbar()
 		self:UpdateName()
@@ -66,8 +67,8 @@ function UF:CreateTargetFrame(frame)
 
 	local health = self:CreateHealth(frame, textParent)
 	health:SetFrameLevel(level + 1)
-	health:SetPoint("LEFT", frame, "LEFT", 0, 0)
-	health:SetPoint("RIGHT", frame, "RIGHT", 0, 0)
+	health:SetPoint("LEFT", frame.Insets.Left, "RIGHT", 0, 0)
+	health:SetPoint("RIGHT", frame.Insets.Right, "LEFT", 0, 0)
 	health:SetPoint("TOP", frame.Insets.Top, "BOTTOM", 0, 0)
 	health:SetPoint("BOTTOM", frame.Insets.Bottom, "TOP", 0, 0)
 	health:SetClipsChildren(true)
@@ -75,12 +76,10 @@ function UF:CreateTargetFrame(frame)
 
 	frame.HealthPrediction = self:CreateHealthPrediction(frame, health, textParent)
 
+	frame.Portrait = self:CreatePortrait(frame)
+
 	local power = self:CreatePower(frame, textParent)
 	power:SetFrameLevel(level + 1)
-	power:SetPoint("LEFT", frame, "LEFT", 0, 0)
-	power:SetPoint("RIGHT", frame, "RIGHT", 0, 0)
-	power:SetPoint("TOP", frame.Insets.Bottom, "TOP", 0, -2)
-	power:SetPoint("BOTTOM", frame.Insets.Bottom, "BOTTOM", 0, 0)
 	frame.Power = power
 
 	power.UpdateContainer = function(_, shouldShow)
@@ -94,6 +93,8 @@ function UF:CreateTargetFrame(frame)
 			end
 		end
 	end
+
+	frame.Insets.Bottom:Capture(power, 0, 0, -2, 0)
 
 	frame.Castbar = self:CreateCastbar(frame)
 
@@ -129,7 +130,6 @@ function UF:CreateTargetFrame(frame)
 
 	local border = E:CreateBorder(textureParent)
 	border:SetTexture("Interface\\AddOns\\ls_UI\\assets\\border-thick")
-	border:SetSize(16)
 	border:SetOffset(-6)
 	frame.Border = border